Lib: only log added after the save if id is db gen
In some cases, the id is generated on the database. This creates a
problem that the id on the logs would be always as zero. So, to avoid
this, we introduced the property IdGeneratedByDatabase. In case of true,
the log for added entities will only be done after the first SaveChanges

Realize, please that it will generate two SaveChanges to database: one
for the entities itself + updated entities' logs + deleted entities' log
and another for the added entities' logs.